requirements.txt
python-telegram-bot==12.0.0b1
google-api-core==1.13.0
dialogflow==0.6.0
vk_api==11.5.0
requests==2.21.0
